Have the glob and walk functions track the full path.
Previously the glob and walk functions maintained the current depth but
only provided the callback function with the current object which in
turn only held its own name not that of its parents.  Those programs
wanting to output the full path had to track the full path with the aid
of the depth.  While reliable to walk this might not have been so for
glob.

This change moves the tracking of the full path into the walk and glob
functions and in turn the callback now takes the path instead of the
depth argument.