Fix wildmat handling of '*' immediately before '.'
In a pattern, when '*' occurs as the last item this is considered a
match whatever the rest of the candidate name.  When the '*' instead
occurs immediately before a '.', i.e. as the last character for the this
component of the pathname, it should likewise be considered a match but
only if the current item is a directory.