Store attributes as per OSFILE for ease of .inf files.
When storing Acorn files on another OSes filesystem the Acorn-specific
attributes are often kept in a sidecar (.inf) file.  This in turn stores
the attributes in the same format as returned by OSFILE.  This change
causes the acorn_fs_object to also keep the attributes in a single
unsigned field in the OSFILE format for ease of reading/writing these
sidecar files.