Add ide2scsi and scsi2ide.
These are a pair of utilities for converting hard disk images.

SCSI emulation and projects like BeebSCSI usually store these images
with bytes in the image file corresponding to bytes as seen by the BBC
micro and can work with the 256-byte sectors the BBC micro expects.

IDE interfaces and sometimes SD card interfaces with a 512-byte sector
size often store one 256-byte sector as seen by the BBC micro in a
512-byte sector on the card by padding with zeros between each byte.

These utilities allow that padding to be stripped/added to convert an
image from one storage format to the other.