Fix a bug with directory full not being detected.
Prior to this commit, a full directory was detected when searching for a
filename reached the end of the directory, but the search was often
abandoned early because the next entry in the directory was
alphanumerically after the one being looked for.  In that case a full
directory was not detected and one entry would be lost as the new one
was inserted.

This commit checks that the end of directory zero byte can be found
before the footer before shifting the entries to make space for the new
one.